Bug 508446 - [memory] Preferences dialog empty tree on the left
This apply to the Floating Point and Traditional Renderings, 
When calling these preferences from the drop down inside the memory
view, the associated preferences dialog comes up with the left side
empty, this left side shall reflect the preference page three path.

The call to the preference dialog has been replaced to use
PreferenceUtil.createPreferenceDialogOn  which takes care of this.

Change-Id: I89d6b1289a7eac5ffcca54a1ca6d8b3c3ed831df