Bug 507336 - Dynamic Printf parameters format may not be clear
The Properties dialog displayed for a Dynamic printf brakpoint
display a "printf ("  field, however the end parenthesis is not added
and not expected as part of the parameter list as the ending parenthesis
is added internally.
  If the user adds it to complete the opening parenthsis the printf does
not work.

This update adds a validation to this field so if the user adds a final
parenthesis, the dialog will display and error message and the OK button
will be disabled.

Change-Id: Ib25c5a11176c6f92c7367246483be0f703e145bd