Update patch set 2
Patch Set 2:

> Thanks for the review.  

My pleasure.

 > Can you clarify what you mean by unexpected
 > failed renderings 

This was showing ???????? instead of real data values.

> and maybe give me steps to reproduce it?

That will be a bit harder, it was random and my guess is that the renderings may have been related to the Change in 84701 that is intermixed. The random failures happened when quickly changing contexts (clicking quickly around the Debug view).

HTH

Patch-set: 2