Update patch set 1
Patch Set 1:

Yeah, I must have just imagined this working, because it couldn't have been. IndexerASTVisitor doesn't compute an enclosing definition for macro reference names, because macro reference names are not encountered during a normal AST traversal, so the technique of "push a function or class definition onto the stack when we enter it" (which is used to compute enclosing definitions for normal names) doesn't work.

Patch-set: 1
Label: Verified=0