Bug 508254 - Static field of enclosing class type
We previously guarded against recursion in this case, but the
result was an invalid composite value that could cause problems
down the line.

This patch avoids getting into the recursion to begin with by
skipping the processing of static fields in a case where it's
not necessary to begin with.

Change-Id: Ic3a346092bb7ad5c94cd15871110dd17ecd64886