Update patch set 2
Patch Set 2:

> I've removed the code for change 84701.  I wasn't sure how to
 > update this review so I've created a new gerrit review here:
 > 
 > https://git.eclipse.org/r/#/c/86958/

Hi Norman,

No problem a normal problem. Please set the Change-Id: of the commit to the same as this review and push again. It is the "magic" Change-Id that makes gerrit put them together. Then please abandon the duplicate!

If you are having trouble with the tooling, let me know and I can help further.

Jonah

Patch-set: 2