Update patch set 2
Patch Set 2: Code-Review-1

(4 comments)

What change is supposed to be seen with this patch?
What I see is that the Debug view properly shows the thread as suspended instead of showing it as running (after the -exec-continue failed).
With the old console, we also see an error message which can help the user know what to fix, since stepping and resuming don't work at all while the breakpoint at *0x0 remains.

However, as Mikhail pointed out on the CDT call, with the new GDB console, we don't see that error message in the console, so the session seems to be hopelessly stuck.  This may be a bug in GDB... I'll have to check.  But I wondered if we should pop-up a dialog in case of such an error?

I've also noticed a couple of other corner cases that I thought I'd mention, although they are not blockers:

1- when running the old GDB console (say with GDB 7.11), once I've set the bp at *0x0, if instead of pressing the resume button, I type 'continue' in the console, the DV goes back to wrongly showing the thread as running.  I believe it is because MIAsyncErrorProcessor checks that we send an MICommand, when in this case, we are sending a CLI command.

2- when running the new GDB console (GDB 7.12 on Linux), once I've set the bp at *0x0, if instead of pressing the resume button, I type 'continue' in the console, the DV goes back to wrongly showing the thread as running. This is because the error is only reported in the console and not on the MI channel at all!  There is nothing we can do about that right now.  I will report this to GDB as I feel this is a limitation in the new console implementation.

Patch-set: 2
Reviewer: Gerrit User 457 <457@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=-1