Update patch set 2
Patch Set 2:

This patch is intended to handle asynchronous errors in GDB. This type of errors appear when an "exec-*" command is successfully issued but something is wrong, the command is aborted and GDB is suspended. Here is the MI output:

842,966 [MI]  40-exec-continue --thread 1
842,967 [MI]  40^running
842,969 [MI]  *running,thread-id="1"
842,969 [MI]  &"Warning:\n"
842,971 [MI]  &"Cannot insert breakpoint 2.\n"
842,972 [MI]  &"Cannot access memory at address 0x0\n"
842,972 [MI]  &"\n"
842,972 [MI]  40^error,msg="Command aborted."
842,978 [MI]  (gdb) 

With this patch the corresponding thread will be shown as suspended. The thread's label will include the message from the error event which in this case is "Command aborted". The message is not informative at all but at least thread is not shown as "running" when it is suspended.

I can try storing the log output for the latest "exec-*" command and use a pop-up to show it to the user but I'm not sure it would work in some unknown cases - there is no way to separate the relevant log messages from irrelevant.
Do you think it would be better? Or we should just try fixing GDB to show a better message instead of "Command aborted".
In CDI we had a special error mar in the image for threads that are suspended because of an error.
 
I can also look into the CLI command case.

Patch-set: 2
Label: Verified=0