Update patch set 2
Patch Set 2:

> With this patch the corresponding thread will be shown as suspended.
> The thread's label will include the message from the error event
> which in this case is "Command aborted". The message is not
> informative at all but at least thread is not shown as "running"
> when it is suspended.

It is better than what we have now.

> I can try storing the log output for the latest "exec-*" command
> and use a pop-up to show it to the user but I'm not sure it would
> work in some unknown cases - there is no way to separate the
> relevant log messages from irrelevant.

That is probably a larger effort, of showing errors to the user
more appropriately.  If we could end up with dialogues popping-up
too often, that would be quite bad.

> Do you think it would be better? Or we should just try fixing GDB
> to show a better message instead of "Command aborted".

Maybe you can generate the error message yourself, when you find
one that says "Command aborted"?
Personally, I didn't even notice this error in the DV, since
the thread simply stayed suspended.  But again, it is better
than what we have now.

> In CDI we had a special error mar in the image for threads that
> are suspended because of an error.

That is nice, but to be honest, I wonder if it will be more noticeable?
Probably not worth it for this special rare case.

So, in the end, this patch is still a step in the right direction,
so we might as well go ahead with it.  But there are a couple of
in-line comments that could be addressed.

Patch-set: 2