Update patch set 2
Patch Set 2: Code-Review+1 Verified+1

Looks good to me. The formatting seems stable (ie issue two formats leads to no changes). Although there is a strange case that had me concerned it may not be stable, but I don't think you actually introduced it. I point it out here for your consideration (especially as you have been diving into that code).

With default K&R formatter, these do not change:

    Tpl<Tpl<Tpl<int>>> tpl3_int;
    Tpl<Tpl<Tpl<int> > > tpl3_int;
    Tpl<Tpl<Tpl<int> >> tpl3_int;

But this:

    Tpl<Tpl<Tpl<int>> > tpl3_int;

formats to:

    Tpl<Tpl<Tpl<int>>> tpl3_int;

What seems to happen is if two > are touching it brings the next one in too, a larger example:

    Tpl<Tpl<Tpl<Tpl<Tpl<Tpl<int> > >> > > tpl6_int;

formats as:

    Tpl<Tpl<Tpl<Tpl<Tpl<Tpl<int> > >>> > tpl6_int;

but I would have expected this. 

    Tpl<Tpl<Tpl<Tpl<Tpl<Tpl<int> > > > > > tpl6_int;

or

    Tpl<Tpl<Tpl<Tpl<Tpl<Tpl<int>>>>>> tpl6_int;


It seems to be a "disagreement" in the implementation between white space before and after closing bracket. If both before and after are don't insert, it always removes the spaces and if both are do insert, then it always adds all the spaces.

I suspect it is all normal confusion between shift and greater-than, but I thought I would point it out, especially as you have just been looking at this code.

Patch-set: 2
Label: Verified=+1
Label: Code-Review=+1