Update patch set 2
Patch Set 2:

I had another look at the code for getting 

 > > But this:
 > >
 > > Tpl<Tpl<Tpl<int>> > tpl3_int;
 > >
 > > formats to:
 > >
 > > Tpl<Tpl<Tpl<int>>> tpl3_int;
 > >
 > > What seems to happen is if two > are touching it brings the next
 > > one in too, a larger example:
 > >
 > > Tpl<Tpl<Tpl<Tpl<Tpl<Tpl<int> > >> > > tpl6_int;
 > >
 > > formats as:
 > >
 > > Tpl<Tpl<Tpl<Tpl<Tpl<Tpl<int> > >>> > tpl6_int;
 > >
 > > but I would have expected this.
 > >
 > > Tpl<Tpl<Tpl<Tpl<Tpl<Tpl<int> > > > > > tpl6_int;
 > >
 > > or
 > >
 > > Tpl<Tpl<Tpl<Tpl<Tpl<Tpl<int>>>>>> tpl6_int;
 > >
 > >
 > > It seems to be a "disagreement" in the implementation between
 > white
 > > space before and after closing bracket. If both before and after
 > > are don't insert, it always removes the spaces and if both are do
 > > insert, then it always adds all the spaces.
 > >
 > > I suspect it is all normal confusion between shift and
 > > greater-than, but I thought I would point it out, especially as
 > you
 > > have just been looking at this code.
 > 
 > Oh, I think I caused this behavior. I'll look into that and update
 > the code, thanks!

I had another look at the code and added your cases in a test. I figured out that the whole handling of template ids can be simplified. Counting of nesting level is not neccessary. Just the lookahead to see whether there are two consecutive closing angle bracksts (i.e. IToken.tSHIFTR) is next. If so only an additional space might be required. The default handling of the visitor will then write the first closing angle bracket for the inner template id (only the first). The surrounding template id is treated similarly: The next token is still the IToken.tSHIFTR. If a space is required due to the configuration it is added. The angle bracket is written by the default handling. Got rid of some code and dumped my changes to the scribe.

Jonah, I hope you don't mind having another look at it.

Patch-set: 2