Update patch set 3
Patch Set 3: Verified-1

(21 comments)

Hi Norman,

Here are my notes on this change.

I have one outstanding issue, but I don't know what to do about it here, so I am going to open it out to the wider community. That is where to save the settings. At the moment you are saving them in the launch configuration, but I believe there has been some chatter in the past about how that is a problem. There is at least one outstanding bug because the other memory view does save in the launch configuration. See bug 500988.

To resolve this, in the first instance I am going to ask Doug to comment as he has most recently raised the issue, but we might go to cdt-dev for this.

Jonah

Patch-set: 3
Label: Verified=-1