Bug 303808: Update ConsolePageParticipant for GDB process
This patch officially limits the ConsolePageParticipant to handling
consoles from the platform's console view, and not any consoles from the
new Debugger Console view.  The code currently behaves like this, but
this patch cleans it up and no longer even checks for GDB consoles.

The ConsolePageParticipant used to provide two things for the GDB
console, when it was part of the platform's Console view:
1- it would select the GDB console when the GDB process was selected
2- it would provide a 'save' button for the GDB console

For #1, this patch selects the first inferior console when the GDB
process is selected. The synchronization of the Debugger Console view is
already handled by each console itself.

For #2 the save button must be handled differently based on the
different type of debugger console (basic vs full).  This should be done
in another commit.  Besides, the code, before this patch, was not adding
the save button to the debugger consoles since those consoles are not an
instanceof org.eclipse.debug.ui.console.IConsole as first checked in
ConsolePagePartipant#isConsoleGdbCli()

Change-Id: I0813ad2261633ce8630ab5842f06a047868856e8