Bug 303808: Make it easier to extend Debugger Console pages
To perform actions on the GdbFullCliConsole, it is important to provide
access to the terminal widget.  This patch adds the new method
GdbFullCliConsolePage#getTerminalViewControl() to give access to the
ITerminalViewControl.

Also, to be consistent with the platform's ConsoleView, this patch also
adds the same groups in the toolbar.  Namely
IConsoleConstants.LAUNCH_GROUP and IConsoleConstants.OUTPUT_GROUP.
Extenders can use these groups to add their buttons to the Debugger
Console view.

Change-Id: I217ce54ada63796de776c6b4cb6b925c3feb7eea