Bug 508543 - Reverse Bp Action doesn't display type properly for editing
The value returned by IBreakpointAction#getTypeName() must match the
value provided in the plugin.xml file for that same action.  The two
names are compared in ActionDialog.createDialogArea() to know if the 
proper action has been found.  This is a fragile technique, that should 
eventually be improved.  For now, we just fix the current problem.

Change-Id: I70fd8b90fc680d3a5d55a95bdd5193d382b6c6f2