Bug 507950 - Answer query on MI channel to avoid GDB waiting forever
With GDB 7.12, it is possible to receive queries on the dedicated MI
channel.  This channel is not accessible or shown to the user so if we
don't answer, GDB will wait forever.

This patch blindly answers 'y' to any query on the MI channel unless it
has already been answered automatically (which happens when we don't use
the full console).

Change-Id: I0e208fc3495ce6ba57b3e477661f47e50680fd88