Bug 510885 Fix case where two env vars try to prepend
The env var operation processer always turned the resulting operation
into a replace, even if it was two prepends. That kills the actual
prepend operation. In this case, keep it as a prepend. Also fix the
append case this way.

Unfortunately this is still broken in other multiple operation
cases, like appending then prepending.

Change-Id: Ib158317f16d025a479201745159a2aa3c1335e25