Update patch set 1
Patch Set 1:

Very nice change!
I am wondering if it's necessary to create a new MIRunControlEventProcessor_7_12 or would it be better to keep the 7_0 version complete, as this is an addition, it should not interfere when using older GDB's. 
  The other advantage is that we could even directly cherry-pick this fix to CDT 9.2.1

In any case it is a go from my side.
Thanks !

Patch-set: 1