Update patch set 1
Patch Set 1:

Salut Vincent,

 > Good for me. No test to add ?
Did you have a chance to try this on a Windows machine? 

Re test: That would be ideal. We've played a bit with SWTBot, but so far we do not have automated UI tests for CDT. So the testing has to be manual for this bit.

Patch-set: 1