Update patch set 1
Patch Set 1:

Works on Mac as intended in the commit message. But I can't really make an educated review of the code or solution proposed here. It's not clear to me if this field should be a path or a command. In the case of a path, there shouldn't need to be any quoting/escaping as there is no ambiguity with the space separating the command and the arguments.

Patch-set: 1