Bug 510987 - Properly support PIE executables as build output
When checking whether gathering the build output from a
CBuildConfiguration, if a binary is a shared library and is an
ELF file, dig deeper and see if it has an INTERP section in the
program header. That tells us it's actually a PIE executable and
add it to the list.

As a rider to this bill, improve the error message when a Qt install
can't be found for a given toolchain/target.

Change-Id: I8fd0bf2d204e9425b02916d7b17f4309a5ad9dd5