Update patch set 1
Patch Set 1:

> I am wondering if it's necessary to create a new MIRunControlEventProcessor_7_12
 > or would it be better to keep the 7_0 version complete, as this is
 > an addition, it should not interfere when using older GDB's.

It would work with the 7_0 version, but I find it increases the risk.  There are no situations where we expect this code to need to run for GDB < 7.12, so just in case we missed some detail for older GDBs that may trigger the new code, it is probably safer to not impact their code at all.

 > The other advantage is that we could even directly cherry-pick this
 > fix to CDT 9.2.1

Yeah, for the 9_2 branch, we'll have to do that to avoid adding APIs.  So the risk will be slightly higher, but that code will only live in a single version and then be replaced by 9.3.

Patch-set: 1