Update patch set 1
Patch Set 1: Code-Review+2 Verified+1

> Just for curiosity,
 > you mentioned that CDT answers automatically when not using the new
 > Full GDB console, where is this being answered ?  Just for my own
 > reference.

It is actually GDB that answers itself when it notices that the MI channel is not using a PTY.  You can see that by running the basic console (GDB 7.11 for example) and typing 'kill' in the console.  If the gdb traces you will see:

949,396 50-interpreter-exec console kill
949,397 ~"Kill the program being debugged? (y or n) [answered Y; input not from terminal]\n"

This is why we never had to worry about queries before.
With the full console though, we open the MI channel using a PTY, so GDB thinks it is talking to a terminal and does not answer the query.  To me, this is a bug in GDB which should know better than to query on a dedicated MI channel (especially since there is a CLI channel running, so obviously the user is not on the MI channel).

Patch-set: 1
Label: Verified=+1
Label: Code-Review=+2