Update patch set 2
Patch Set 2:

> Build Unstable
 > 
 > https://hudson.eclipse.org/cdt/job/cdt-verify/7243/ : UNSTABLE

There are three failures here.  The first and third we can ignore for now and will pay attention after a rebase I'm about to do.

The second failure is what I was looking for.  It is the failure of the new test that I added.  I was supposed to fail since this patch is based on a commit preceding the fix we're testing.  The test output is:

org.eclipse.cdt.tests.dsf.gdb.tests.MIRunControlReverseTest.testQueryHandling[gdb.7.12]
Error Message
Timed-out waiting to write to a variable, probably because of a query.

This confirms that the test will trigger if a query is not answered.

I will now rebase this patch on top of the commit that fixes the problem and the test should no longer fail.

Patch-set: 2
Label: Verified=0