Update patch set 6
Patch Set 6: Code-Review-2

> I have tested the upgrade and I've noticed those were not updated
 > or replaced by the new ones:
 > 
 > C/C++ C99 LR Parser Source
 > C/C++ IBM XLC Compiler Support Source
 > C/C++ UPC (Unified Parallel C) Support Source

I see in the upgrade dialogue that p2 does not show an upgrade but a new installation for the features replacing these.  However, once done, when looking at the installation details, I see the new features and not the old ones.  How did you reproduce this?

 > Was that a known limitation we discussed? Or perhaps they need
 > p2.inf?

Looking at the old ID of these source features, they do have changed, so you are right, we need a p2.inf.

What might be happening in my test is that I install the SDK for these features, and the SDK gets properly replaced, taking the old features out with it.  But the source feature themselves would not upgrade on their own.  They weren't in the category.xml file though, so I don't quite know how they could be installed by themselves...

Patch-set: 6
Label: Code-Review=-2