Prototype taking-over an existing GDB session, started outside Eclipse.
This is on top of the "FLUID session" prototype. 

# Start a GDB (7.12 +) outside Eclipse, ideally using a binary that is
part of a CDT project/workspace, and had been compiled with debug
symbols : 
$> gdb.7.12.1 /path/to/binary

a) setup all-stop mode, for example:
set mi-async on
set pagination off
# start then interrupt the program
r&
interrupt

b) setup non-stop mode, for example:
set mi-async on
set pagination off
set non-stop on
# start then interrupt the program
r&
interrupt -a


# *** in CDT, make sure that the same GDB mode (all-stop/non-stop) is
selected as the default in the GDB preferences (Window -> Preferences ->
C/C++ -> Debug -> GDB -> "Non-stop mode" checkbox. 
Also make sure that, in the same page, the default "GDB Debugger" points
towards a GDB of the same version as the one you're using.

(note: ideally we would automatically detect the non-stop/all-stop mode
and GDB version automatically. However it's not trivial, given that we
need to know the version and mode to instantiate the correct services,
and we need the correct services running to interact with GDB. Using the
preferences defaults is an easy work-around for now)

# To take-over the debug session, in CDT, use the "Take-over external
GDB Session" button, located on the toolbar of the Debugger Console. 

# Look for a trace like this in the development Eclipse console: 
MI PTY: /dev/pts/37

# In GDB, "connect" to the PTYs created by CDT, from the printout above: 
new-ui mi2 /dev/pts/37


# after a pause of ~30 seconds to give you time to configure the PTYs,
CDT will continue execution and should use the external  console as if
it were the built-in CDT Debugger console. 

Notes:
- upon initially taking-over the debug session, the state of the threads
in CDT may not reflect GDB's state. You can refresh the Debug View or
start/stop the binary to get back in a correct state.
- adding a new process to the debug session from CDT (i.e. "Debug New
Executable") doesn't seem to work; a new inferior is created in GDB but
not properly loaded/started. Attaching seems to work.
- at some point the Debugger Console became disabled, even for normal
launches. No time to investigate why yet...



Change-Id: Ia781629b3a5ef4199e10239c334188d97cc60665