Update patch set 2
Patch Set 2:

> I don't quite understand your point..
 > 
 > partOpened() isn't doing anything.

Right.  I was awkwardly trying to say that if the class (PinDebugContextActionDelegate.java)
was not disposed, it maybe continue to be used somehow (I don't know if that is possible).
I just used partOpened() as one possibility of the platform saying "the view has been
re-opened".  In that case, PinDebugContextActionDelegate, before this patch, would still be in
a valid state.  However, after this patch, it is not in a valid state because we force
a dispose().


 > Also, from my understanding,
 > after partClosed() is notified no other part*() notifications will
 > occur for that part, since it is gone forever. Therefore I don't
 > understand the "If there could be a scenario where partClosed() is
 > called but dispose() is not, then partOpened() could be called" ...

I was just playing it safe.  If the original code treated partClosed()
differently than dispose(), I figure it might be for such a scenario.
So, like the code does now, let's have partClosed() only setCheck(false)
and removeDebugEventFilter(fPart) instead of calling the full dispose.

The fix should then be to have dispose also call setCheck(false) and
removeDebugEventFilter(fPart).

 > The reference to a "race" is also not clear to me, do you mean a
 > data race? But there is no multi-threading here.

Sorry, I meant "race condition" where dispose() can get called before
partClosed(), so dispose() must perform the same action as partClosed().

Patch-set: 2
Label: Verified=0