Bug 399494 - Add test for variable update of pointers declared with typedef
This patch adds a test to check that the value of a pointer declared
using a typedef will update correctly in CDT as it changes in the
program.  There are two distinct scenarios, pointers that are variables
and pointers that are fields of structures.

Instead of adding content to testUpdateOfPointer, I decided to make a
similar test method, testUpdateOfPointerTypedef.  The original test
method was getting too long and was difficult to follow.  I think it's
good to keep them short and focused.

The test is failing currently, since the fix is in the following
patch.  I added an @Ignore annotation which is removed along with the
fix.

Change-Id: Id9ac07d7fbbe76e991c77c1260a1b5df132c639f
Signed-off-by: Simon Marchi <simon.marchi@polymtl.ca>