Bug 399494 - Consider all variable objects as potentially modifiable
There are cases where we consider some variable objects as
non-modifiable even though they are.  In particular, if a pointer is
declared using a typedef, isModifiable will return false.  This is
because it has a child (the pointed value), but CDT doesn't know it's a
pointer.

Since there's no easy way to know for sure whether a variable object is
modifiable, just assume everything is.  The methods referring to
modifiable descendants are deprecated and replaced with equivalents
without the "modifiable" qualifier:

  - addModifiableDescendant -> addDescendant
  - removeModifiableDescendant -> removeModifiableDescendant

The isModifiable method is also deprecated and now simply returns true,
so that any variable object is considered modifiable.

Change-Id: If43b3b6e49cd4a20ea929c2a096745a32de14cd0
Signed-off-by: Simon Marchi <simon.marchi@polymtl.ca>