Update patch set 2
Patch Set 2:

(7 comments)

Thanks for writing these quick fixes!

It would be great to have tests for them. We have existing tests for quick fixes for codan markers, but not for quick fixes for compiler error markers. How involved would it be to create compiler error markers in a test suite, so quick fixes for them could be tested?

Patch-set: 2
Label: Verified=0