NewLaunchConfig* notify listeners when table entries are selected
The NewLaunchConfig*Page classes don't notify their listeners when they
automatically select a type or mode with table.select().  This causes
the NewLaunchConfigEditPage to have a "null" type if the user doesn't
explicitly click on one and thus opens the default page.

Added in calls to table.notifyListeners() in order to fix this.

Verified that the launch config edit page opens appropriately even if
there is no explicit selection by the user

Change-Id: I5fbafd2f43d7c0951e929f5dcbfc57fcccd18db6
Signed-off-by: Matthew Bastien <bastien.matthew@gmail.com>