Remote debug: Use the IFileStore interface to force the file executable
Instead of calling "chmod +x" by hand, use IFileStore.putInfo to set the
file attributes.  Since we already use the IFileStore API to copy the
file, it makes sense (and it's cleaner) to use it to set the executable
attribute as well.

In most cases, it wouldn't be needed to make the file executable, since
it should already be on the host filesystem and IFileStore.copy
transfers the attributes.  However, it's still good to force it
executable in case it's not already for some reason.

Change-Id: I4c86e36265962781d4541aaceeb40b502248f674
Signed-off-by: Simon Marchi <simon.marchi@polymtl.ca>