Remote debug: Do not manually chmod +x the file after transfer
While stepping in the remote debug code, I noticed that the copy method
of Eclipse Filesystem preserves file attributes, including permissions.
This is what is used in the RemoteHelper.remoteFileDownload method used
to send the binary to the target.  This means that we don't need to
execute chmod +x manually after the file transfer to make it executable,
since it should already be executable on the host.

The transfer of the file attributes is done here:

	JschFileStore.putInfo(IFileInfo, int, IProgressMonitor) line: 319
	LocalFile(FileStore).transferAttributes(IFileInfo, IFileStore) line: 465
	LocalFile(FileStore).copyFile(IFileInfo, IFileStore, int, IProgressMonitor) line: 201
	LocalFile(FileStore).copy(IFileStore, int, IProgressMonitor) line: 130
	LocalFile.copy(IFileStore, int, IProgressMonitor) line: 124
	RemoteHelper.remoteFileDownload(ILaunchConfiguration, ILaunch, String, String, IProgressMonitor) line: 116
	RemoteGdbLaunchDelegate.launch(ILaunchConfiguration, String, ILaunch, IProgressMonitor) line: 61
	LaunchConfiguration.launch(String, IProgressMonitor, boolean, boolean) line: 885
	LaunchConfiguration.launch(String, IProgressMonitor, boolean) line: 739
	DebugUIPlugin.buildAndLaunch(ILaunchConfiguration, String, IProgressMonitor) line: 1039
	DebugUIPlugin$8.run(IProgressMonitor) line: 1256
	Worker.run() line: 56

If we still want to force the file to be executable, because for some
reason it might not be already executable on the development machine, we
should probably do that through the FileStore interface
(FileStore.putInfo) instead of calling chmod directly.

Change-Id: I4c86e36265962781d4541aaceeb40b502248f674
Signed-off-by: Simon Marchi <simon.marchi@polymtl.ca>