Update patch set 6
Patch Set 6:

Marc, I have submitted this, it looks good to me. Auto-testing of this type of UI is very difficult, I don't know how I would construct such a test case that adds value. Best to rely on the spreadsheets in this case and I see that Debugger Console has a line in there already.

Patch-set: 6