Update patch set 3
Patch Set 3:

(14 comments)

Very nice to see this enhancement, 
It would be better if we could actually have a default implementation included.
  What about using the name of the executable as the serialization id? since the executable name includes the full location path these could provide distinct names for 32bit and 64bit versions of the same binary.

if the name of the executable can actually be the same for two different programs, then the behavior would be equivalent to what we have without this enhancement.

There is an example on how to get the process name in the following test case.
GDBProcessesTest#getProcessData 

Let me know what you think,

I have provided patch set 4, which addresses most of the inline comments, feel free to remove what you disagree with so we can discuss them on a new patch set.

Thanks !

Patch-set: 3
Label: Verified=0